requirements.txt
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0.0
